refactor: remove deprecated server and database configuration files, and implement startServer function in index.ts
