+++ src/config/index.ts
@@ -6,7 +6,6 @@ import './dotenv';
 export * from './database';
 export * from './server';
 export * from './websocket';
-
 /**
  * Configuration for the server.
  */
